backend/requirements.txt
fastapi
uvicorn
pymongo
requests
werkzeug
python-dotenv
python-jose 
google-auth
passlib==1.7.4 
bcrypt==3.2.2
